pyproject.toml
[tool.poetry]
name = "weatherapp"
version = "0.1.0"
description = "Retrieves and publishes weather data"
authors = ["Justin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pandas = "^2.0.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
